Shader great rework in progress

+++ src/parser/tools.c
@@ -6,7 +6,7 @@
 /*   By: gpinchon <gpinchon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/19 01:27:17 by gpinchon          #+#    #+#             */
-/*   Updated: 2017/02/23 23:16:38 by gpinchon         ###   ########.fr       */
+/*   Updated: 2017/03/26 23:41:02 by gpinchon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,9 +38,9 @@ char	**split_path(const char *path)
 	return (final_path);
 }
 
-unsigned long hash(unsigned char *str)
+long long int hash(unsigned char *str)
 {
-    unsigned long hash = 5381;
+    long long int hash = 5381;
     int c;
 
     while ((c = *str++))
